requirements-full.txt
absl-py==2.4.0
accelerate==0.22.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
decord==0.6.0
diffusers==0.37.0
einops==0.8.2
fastapi==0.135.1
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
future==1.0.0
gradio==3.36.1
gradio_client==2.4.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
imageio==2.27.0
imageio-ffmpeg==0.4.9
importlib_metadata==9.0.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
linkify-it-py==2.1.0
markdown-it-py==2.2.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
narwhals==2.18.1
networkx==3.6.1
numpy==1.26.4
nvidia-cublas==13.0.0.19
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.48
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.48
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.0.42
nvidia-cufile-cu12==1.13.1.3
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.3.29
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.2.49
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.21.5
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==12.1.1
propcache==0.4.1
protobuf==4.25.9
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.17.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
starlette==0.52.1
sympy==1.14.0
tokenizers==0.22.2
torch==2.9.0+cu130
torchaudio==2.9.0+cu130
torchvision==0.24.0+cu130
tqdm==4.67.3
transformers==5.3.0
triton==3.5.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
yarl==1.23.0
zipp==3.23.0
